Moved files to folder molgenis_emx2_client
